Ported changes from ZDoom revision 1748-1760 and from GZDoom revision 410-413: - Added the MF6_STEPMISSILE flag so that the Whirlwind can "walk" up steps. - Changed the dword definition of PalEntry to uint32 so that it has one consistent definition across all source files. - Swapped the order of floor and ceiling moves in DElevator::Tick() so that if an elevator contains an actor exactly the same height as it, it will not be blocked. - Fixed: FWeaponSlot::PickWeapon() wrapped around improperly when the starting value for i was 0. - Added kgsws's SummonActor enhancement and bumped netgame and demo versions because this submission changes an existing command. - fixed some overzealous Linux warning removal resulting in broken code. - Added player MugShotMaxHealth property. Negative values use the player's max health as the mug shot max health, zero uses 100 as the mug shot max health, and positive values used directly as the mug shot max health. - Added buddha cheat. - Added TELEFRAG_DAMAGE constant, and changed the two places that still used 1000 as the threshold for god mode damage to use it instead. (Players with MF2_INVULNERABLE set already used 1000000 as their threshold.) - Added MF6_NOTELEFRAG flag. - Fixed: M_QuitResponse() tried to play a sound even when none was specified in the gameinfo. - Added Yes/No selections for Y/N messages so that you can answer them entirely with a joystick. - Fixed: Starting the menu at the title screen with a key other than Escape left the top level menu out of the menu stack. - Changed the save menu so that cancelling input of a new save name only deactivates that control and does not completely close the menus. - Fixed "any key" messages to override input to menus hidden beneath them and to work with joysticks. - Removed the input parameter from M_StartMessage and the corresponding messageNeedsInput global, because it was redundant. Any messages that want a Y/N response also supply a callback, and messages that don't care which key you press don't supply a callback. - Changed MKEY_Back so that it cancels out of text entry fields before backing to the previous menu, which it already did for the keyboard. - Changed the menu responder so that key downs always produce results, regardless of whether or not an equivalent key is already down. - Added support for per-SpawnShot spawn lists. (Also fixed a potential infinite loop, though I'm not sure the parser allowed this condition to happen.) - Changed the DWORDs in dobject.h into uint32s, since they were preventing edit-and-continue from working for the Windows source files. - When a WM_KEYDOWN message is received with VK_PROCESSKEY, the scan key is now used to retrieve the real virtual key for the message. This fixes the previous issue that caused me to completely disable the IME. - Removed the code that disables the IME, since it also disables the ability to switch between keyboard layouts that do not use an IME. - TranslateMessage() is no longer called if GUI capture mode is off, so no dead key processing is performed until it might be useful.

+++ src/gl/old_renderer/gl1_flats.cpp
@@ -581,13 +581,13 @@ void GLFlat::ProcessSector(sector_t * frontsector, subsector_t * sub)
 			fixed_t lastceilingheight=sector->CenterCeiling();	// render only in the range of the
 			fixed_t lastfloorheight=sector->CenterFloor();		// current sector part (if applicable)
 			F3DFloor * rover;	
-			unsigned int k;
+			int k;
 			
 			// floors are ordered now top to bottom so scanning the list for the best match
 			// is no longer necessary.
 
 			ceiling=true;
-			for(k=0;k<x.ffloors.Size();k++)
+			for(k=0;k<(int)x.ffloors.Size();k++)
 			{
 				rover=x.ffloors[k];
